multi-db is almost done

+++ api/src/db/bills.rs
@@ -10,7 +10,7 @@ pub struct Model {
     pub id: i32,
     pub owner: String,
     pub image: i32,
-    pub status: i32,
+    pub status: i8,
     pub price: Option<i32>,
     pub reason: Option<String>,
     pub handled_by: Option<String>,
